refactor: FarePrincipalFinder Bean -> Util 로 변경
